+++ .gitignore
@@ -6,3 +6,4 @@ node_modules/
 *.tar.gz
 *.zip
 .idea
+composer.lock
