+++ .travis.yml
@@ -1,5 +1,3 @@
-sudo: false
-
 language: php
 
 notifications:
@@ -14,11 +12,9 @@ branches:
 
 cache:
   directories:
-    - vendor
     - $HOME/.composer/cache
 
 matrix:
-  include:
   include:
     - php: 5.5
       env: WP_VERSION=latest WP_MULTISITE=0 PHP_UNIT_VERSION=4.8
@@ -57,8 +53,3 @@ script:
 
 after_success:
   - travis_retry php coveralls.phar -v
-
-cache:
-  directories:
-  - vendor
-  - $HOME/.composer/cac
